refactor analytics to support question level analytics
