+++ .vscode/snippets/typescriptreact.json
@@ -3,9 +3,7 @@
 		"prefix": "template-sfc",
 		"body": [
 			"import * as React from 'react';",
-			"import { StyledComponentProps } from 'types/component';",
-			"import { withStyles, classNames } from 'styles/jss';",
-			"import { JSSStyles } from 'styles/jss';",
+			"import { JSSStyles, withStyles, classNames } from 'styles/jss';",
 			"",
 			"const styles: JSSStyles = {",
 			"  ${TM_FILENAME_BASE}: {",
@@ -20,7 +18,7 @@
 			"/**",
 			" * ${TM_FILENAME_BASE} React Stateless Component",
 			" */",
-			"export const ${TM_FILENAME_BASE}: = withStyles<${TM_FILENAME_BASE}Props>(styles)(({",
+			"export const ${TM_FILENAME_BASE} = withStyles<${TM_FILENAME_BASE}Props>(styles)(({",
 			"  className, classes, children,",
 			"}) => {",
 			"  return (",
