+++ src/editors/content/question/checkallthatapply/CheckAllThatApply.controller.ts
@@ -1,15 +1,19 @@
 import { connect } from 'react-redux';
 import * as contentTypes from 'data/contentTypes';
-import { QuestionProps } from '../question/Question';
+import { OwnQuestionProps } from '../question/Question';
 import { CheckAllThatApply } from './CheckAllThatApply';
 import { CombinationsMap } from 'types/combinations';
 import { computeCombinations } from 'actions/choices';
 import { toggleAdvancedScoring } from 'actions/questionEditor';
 import { State } from 'reducers';
+import { AnalyticsState } from 'reducers/analytics';
+import { AssessmentModel } from 'data/models';
 
 interface StateProps {
   advancedScoringInitialized: boolean;
   advancedScoring: boolean;
+  analytics: AnalyticsState;
+  assessmentId: string;
 }
 
 interface DispatchProps {
@@ -17,7 +21,7 @@ interface DispatchProps {
   onToggleAdvancedScoring: (id: string, value?: boolean) => void;
 }
 
-interface OwnProps extends QuestionProps<contentTypes.MultipleChoice> {
+interface OwnProps extends OwnQuestionProps<contentTypes.MultipleChoice> {
 
 }
 
@@ -25,6 +29,9 @@ const mapStateToProps = (state: State, ownProps: OwnProps): StateProps => {
   return {
     advancedScoringInitialized: state.questionEditor.hasIn(['scoring', ownProps.model.guid]),
     advancedScoring: state.questionEditor.getIn(['scoring', ownProps.model.guid]),
+    analytics: state.analytics,
+    // this line assumes this component is only used within an assessment and document is loaded
+    assessmentId: (state.documents.first().document.model as AssessmentModel).resource.id,
   };
 };
