+++ src/editors/content/question/draganddrop/DynaDropInput.controller.ts
@@ -1,23 +1,27 @@
 import { connect } from 'react-redux';
 import * as contentTypes from 'data/contentTypes';
-import { QuestionProps } from '../question/Question';
-import { DynaDropInput } from './DynaDropInput';
+import { QuestionProps, OwnQuestionProps } from '../question/Question';
+import { DynaDropInput, DynaDropInputProps } from './DynaDropInput';
 import { State } from 'reducers';
 import { ActiveContext } from 'types/active';
 import { toggleAdvancedScoring } from 'actions/questionEditor';
+import { AnalyticsState } from 'reducers/analytics';
+import { AssessmentModel } from 'data/models';
 
 interface StateProps {
   activeContext: ActiveContext;
   selectedInitiator: string;
   advancedScoringInitialized: boolean;
   advancedScoring: boolean;
+  analytics: AnalyticsState;
+  assessmentId: string;
 }
 
 interface DispatchProps {
   onToggleAdvancedScoring: (id: string, value?: boolean) => void;
 }
 
-interface OwnProps extends QuestionProps<contentTypes.QuestionItem> {
+interface OwnProps extends OwnQuestionProps<contentTypes.QuestionItem> {
   onAddItemPart: (item, part, body) => void;
 }
 
@@ -27,6 +31,9 @@ const mapStateToProps = (state: State, ownProps: OwnProps): StateProps => {
     selectedInitiator: state.dynadragdrop.selectedInitiator,
     advancedScoringInitialized: state.questionEditor.hasIn(['scoring', ownProps.model.guid]),
     advancedScoring: state.questionEditor.getIn(['scoring', ownProps.model.guid]),
+    analytics: state.analytics,
+    // this line assumes this component is only used within an assessment and document is loaded
+    assessmentId: (state.documents.first().document.model as AssessmentModel).resource.id,
   };
 };
