+++ src/editors/content/question/multipart/MultipartInput.controller.ts
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import * as contentTypes from 'data/contentTypes';
-import { QuestionProps } from '../question/Question';
+import { OwnQuestionProps, QuestionProps } from '../question/Question';
 import { MultipartInput, PartAddPredicate } from './MultipartInput';
 import { State } from 'reducers';
 import { ActiveContext } from 'types/active';
@@ -8,11 +8,15 @@ import { setActiveItemIdActionAction } from 'actions/inputRef';
 import { Maybe } from 'tsmonad';
 import { RouterState } from 'reducers/router';
 import { clearSearchParam } from 'actions/router';
+import { AnalyticsState } from 'reducers/analytics';
+import { AssessmentModel } from 'data/models';
 
 interface StateProps {
   activeContext: ActiveContext;
   selectedInput: Maybe<string>;
   router: RouterState;
+  analytics: AnalyticsState;
+  assessmentId: string;
 }
 
 interface DispatchProps {
@@ -20,7 +24,7 @@ interface DispatchProps {
   onClearSearchParam: (name) => void;
 }
 
-interface OwnProps extends QuestionProps<contentTypes.QuestionItem> {
+interface OwnProps extends OwnQuestionProps<contentTypes.QuestionItem> {
   canInsertAnotherPart: PartAddPredicate;
   onAddItemPart: (item, part, body) => void;
 }
@@ -30,6 +34,9 @@ const mapStateToProps = (state: State, ownProps: OwnProps): StateProps => {
     activeContext: state.activeContext,
     selectedInput: state.inputRef,
     router: state.router,
+    analytics: state.analytics,
+    // this line assumes this component is only used within an assessment and document is loaded
+    assessmentId: (state.documents.first().document.model as AssessmentModel).resource.id,
   };
 };
