+++ src/editors/content/question/ordering/Ordering.controller.ts
@@ -1,14 +1,18 @@
 import { connect } from 'react-redux';
 import * as contentTypes from 'data/contentTypes';
-import { QuestionProps } from '../question/Question';
+import { OwnQuestionProps } from '../question/Question';
 import { Ordering } from './Ordering';
 import { PermutationsMap } from 'types/combinations';
 import { computePermutations } from 'actions/choices';
 import { toggleAdvancedScoring } from 'actions/questionEditor';
+import { AnalyticsState } from 'reducers/analytics';
+import { AssessmentModel } from 'data/models';
 
 interface StateProps {
   advancedScoringInitialized: boolean;
   advancedScoring: boolean;
+  analytics: AnalyticsState;
+  assessmentId: string;
 }
 
 interface DispatchProps {
@@ -16,7 +20,7 @@ interface DispatchProps {
   onToggleAdvancedScoring: (id: string, value?: boolean) => void;
 }
 
-interface OwnProps extends QuestionProps<contentTypes.Ordering> {
+interface OwnProps extends OwnQuestionProps<contentTypes.Ordering> {
 
 }
 
@@ -24,6 +28,9 @@ const mapStateToProps = (state, ownProps: OwnProps): StateProps => {
   return {
     advancedScoringInitialized: state.questionEditor.hasIn(['scoring', ownProps.model.guid]),
     advancedScoring: state.questionEditor.getIn(['scoring', ownProps.model.guid]),
+    analytics: state.analytics,
+    // this line assumes this component is only used within an assessment and document is loaded
+    assessmentId: (state.documents.first().document.model as AssessmentModel).resource.id,
   };
 };
