+++ src/editors/document/analytics/Analytics.controller.ts
